backend/requirements.txt
flask==2.3.3
werkzeug==2.3.7
flask-cors==4.0.0
pandas==2.2.1
numpy==1.26.4
openpyxl==3.1.2
python-dotenv==1.0.1
langchain==0.1.12
langchain-openai==0.0.8
scikit-learn==1.3.2
gunicorn==21.2.0
python-decouple==3.8
whitenoise==6.6.0 